Add files via upload
Insurance Data Analysis

Problem Statement: 
An insurance agency, ABC Insurance, has a large dataset containing information about their policyholders and claims. They want to perform exploratory data analysis (EDA) on this dataset to gain insights that can help them make better business decisions and improve their operations. 

The agency wants to analyze the different body types and the environment that affect the premium. The disease's effect or the cost of treatment differs depending on the circumstances. For example, a smoker's medical insurance premium may be higher than that of a healthy person, because smokers are more likely to develop chronic diseases. The agency wants to analyze the data to research healthcare premium costs.

Objective: To analyze the dataset that will help to create a model that will predict the cost of medical insurance based on various input features

Domain: Healthcare

Dataset: insurance dataset (insurance.csv)

Dataset Description:

age	Age of the person
sex	Female or Male
BMI	BMI value to estimate an individual's health and fitness condition
children	number of children (1,2,3,4, or 5)

smoker	The person is a smoker or not
region	Specifies the region (northeast, northwest, southeast, southwest)
charges	the amount of insurance